+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog for AWS SDK for JavaScript
-<!--LATEST=2.1056.0-->
+<!--LATEST=2.1057.0-->
 <!--ENTRYINSERT-->
+
+## 2.1057.0
+* feature: ElastiCache: AWS ElastiCache for Redis has added a new Engine Log LogType in LogDelivery feature. You can now publish the Engine Log from your Amazon ElastiCache for Redis clusters to Amazon CloudWatch Logs and Amazon Kinesis Data Firehose.
+* feature: Glue: This SDK release adds support to pass run properties when starting a workflow run
+* feature: LexRuntimeV2: Updates API to latest version.
+* feature: Nimble: Amazon Nimble Studio now supports validation for Launch Profiles. Launch Profiles now report static validation results after create/update to detect errors in network or active directory configuration.
+* feature: Pinpoint: Adds JourneyChannelSettings to WriteJourneyRequest
+* feature: SSM: AWS Systems Manager adds category support for DescribeDocument API
 
 ## 2.1056.0
 * feature: EC2: Hpc6a instances are powered by a third-generation AMD EPYC processors (Milan) delivering all-core turbo frequency of 3.4 GHz